@@ -1,20 +1,17 @@
-﻿using System;
-using System.Collections.ObjectModel;
-using System.Threading.Tasks;
+﻿using System.Collections.ObjectModel;
 
-namespace GenericViewModels.Services
+namespace GenericViewModels.Services;
+
+public interface IItemsService<T>
 {
-    public interface IItemsService<T>
-    {
-        Task RefreshAsync();
+    Task RefreshAsync();
 
-        Task<T> AddOrUpdateAsync(T item);
+    Task<T> AddOrUpdateAsync(T item);
 
-        Task DeleteAsync(T item);
+    Task DeleteAsync(T item);
 
-        ObservableCollection<T> Items { get; }
+    ObservableCollection<T> Items { get; }
 
-        T? SelectedItem { get; set; }
-        event EventHandler<T>? SelectedItemChanged;
-    }
+    T? SelectedItem { get; set; }
+    event EventHandler<T>? SelectedItemChanged;
 }
