@@ -1,54 +1,51 @@
 ﻿using Microsoft.Toolkit.Mvvm.ComponentModel;
-using System;
-using System.Threading;
 
-namespace GenericViewModels.ViewModels
+namespace GenericViewModels.ViewModels;
+
+public abstract class ViewModelBase : ObservableObject
 {
-    public abstract class ViewModelBase : ObservableObject
+    private class StateSetter : IDisposable
     {
-        private class StateSetter : IDisposable
+        private readonly Action _end;
+        public StateSetter(Action start, Action end)
         {
-            private readonly Action _end;
-            public StateSetter(Action start, Action end)
-            {
-                start?.Invoke();
-                _end = end;
-            }
-            public void Dispose() => _end?.Invoke();
+            start?.Invoke();
+            _end = end;
         }
+        public void Dispose() => _end?.Invoke();
+    }
 
-        private int _inProgressCounter = 0;
-        protected void SetInProgress(bool set = true)
+    private int _inProgressCounter = 0;
+    protected void SetInProgress(bool set = true)
+    {
+        if (set)
         {
-            if (set)
-            {
-                Interlocked.Increment(ref _inProgressCounter);
-                OnPropertyChanged(nameof(InProgress));
-            }
-            else
-            {
-                Interlocked.Decrement(ref _inProgressCounter);
-                OnPropertyChanged(nameof(InProgress));
-            }
+            Interlocked.Increment(ref _inProgressCounter);
+            OnPropertyChanged(nameof(InProgress));
         }
-
-        public IDisposable StartInProgress() => 
-            new StateSetter(() => SetInProgress(), () => SetInProgress(false));
-
-        public bool InProgress => _inProgressCounter != 0;
-
-        private bool _hasError;
-        public bool HasError
+        else
         {
-            get => _hasError;
-            set => SetProperty(ref _hasError, value);
+            Interlocked.Decrement(ref _inProgressCounter);
+            OnPropertyChanged(nameof(InProgress));
         }
+    }
 
-        private string? _errorMessage;
-        public string? ErrorMessage
-        {
-            get => _errorMessage;
-            set => SetProperty(ref _errorMessage, value);
-        }
+    public IDisposable StartInProgress() =>
+        new StateSetter(() => SetInProgress(), () => SetInProgress(false));
+
+    public bool InProgress => _inProgressCounter != 0;
+
+    private bool _hasError;
+    public bool HasError
+    {
+        get => _hasError;
+        set => SetProperty(ref _hasError, value);
+    }
+
+    private string? _errorMessage;
+    public string? ErrorMessage
+    {
+        get => _errorMessage;
+        set => SetProperty(ref _errorMessage, value);
     }
 }
