@@ -1,13 +1,4 @@
-﻿using BooksApp.Services;
-using BooksApp.ViewModels;
-using BooksLib.Models;
-using BooksLib.Services;
-using BooksLib.ViewModels;
-using GenericViewModels.Services;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Toolkit.Mvvm.DependencyInjection;
-using Microsoft.UI.Xaml;
+﻿
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
