@@ -1,25 +1,17 @@
-﻿using Microsoft.UI.Xaml.Data;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace BooksApp.Converters;
 
-namespace BooksApp.Converters
+public class StringToBoolConverter : IValueConverter
 {
-    public class StringToBoolConverter : IValueConverter
+    public object Convert(object value, Type targetType, object parameter, string language)
     {
-        public object Convert(object value, Type targetType, object parameter, string language)
-        {
-            if (value.ToString() == "False")
-                return false;
-            else
-                return true;
-        }
+        if (value.ToString() == "False")
+            return false;
+        else
+            return true;
+    }
 
-        public object ConvertBack(object value, Type targetType, object parameter, string language)
-        {
-            throw new NotImplementedException();
-        }
+    public object ConvertBack(object value, Type targetType, object parameter, string language)
+    {
+        throw new NotImplementedException();
     }
 }
