@@ -1,19 +1,8 @@
-﻿using BooksLib.ViewModels;
-using Microsoft.Toolkit.Mvvm.DependencyInjection;
-using Microsoft.UI.Xaml.Controls;
+﻿namespace BooksApp.Views;
 
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
-
-namespace BooksApp.Views
+public sealed partial class BookDetailPage : Page
 {
-    /// <summary>
-    /// An empty page that can be used on its own or navigated to within a Frame.
-    /// </summary>
-    public sealed partial class BookDetailPage : Page
-    {
-        public BookDetailPage() => InitializeComponent();
+    public BookDetailPage() => InitializeComponent();
 
-        public BookDetailViewModel ViewModel { get; } = Ioc.Default.GetRequiredService<BookDetailViewModel>();
-    }
+    public BookDetailViewModel ViewModel { get; } = Ioc.Default.GetRequiredService<BookDetailViewModel>();
 }
